python/auth_client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "grpcio-tools>=1.50.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-client"
version = "0.1.0"
description = "Generated gRPC client for Auth Service"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Organization", email = "dev@yourorg.com"}
]
keywords = ["grpc", "protobuf", "auth", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.50.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/api-contracts"
Repository = "https://github.com/yourorg/api-contracts"
Documentation = "https://github.com/yourorg/api-contracts/tree/main/python/auth_client"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
auth_client = ["py.typed"]
